+++ src/main/jni/device/device.h
@@ -3,7 +3,6 @@
 
 #include <android/log.h>
 
-
 #define DEF_SERIAL          0
 #define USB_TO_SERIAL       1
 #define FREE_TO_SERIAL      2
@@ -18,11 +17,11 @@ int Write_Port(int fd, void * buffer, int size);
 int Read_Port(int fd, void * buffer, int size);
 int Close_Port(int fd);
 //video
-int Open_Video(int port);
-int Set_Video(int fd, int width, int height, int format);
-int Read_Video(int fd, unsigned char * pFrameBuffer, int size);
-int Close_Video(int fd);
-int Check_Format(int fd, unsigned int format);
+long Open_Video(int port);
+int Set_Video(long fd, int width, int height, int format);
+int Read_Video(long fd, unsigned char * pFrameBuffer, int size);
+int Close_Video(long fd);
+int Check_Format(long fd, unsigned int format);
 
 #ifdef __cplusplus
 }
